+++ pythoncz/pages/jobs/parsers/__init__.py
@@ -1,0 +1,6 @@
+from pythoncz.pages.jobs.parsers import (jobscz, pythonorg, remoteok,
+                                         stackoverflowcom, startupjobscz)
+
+
+__all__ = ('jobscz', 'pythonorg', 'remoteok', 'stackoverflowcom',
+           'startupjobscz')
