+++ pythoncz/pages/jobs/parsing.py
@@ -1,0 +1,10 @@
+import re
+
+
+WHITESPACE_RE = re.compile(r'\s+')
+
+
+def normalize_text(text):
+    if text:
+        return WHITESPACE_RE.sub(' ', text.strip())
+    return text
